[CORDA-1941]: Moved some test utils to a re-usable location.
