+++ tools/shell/src/main/kotlin/net/corda/tools/shell/InteractiveShell.kt
@@ -18,6 +18,7 @@ import net.corda.core.internal.*
 import net.corda.core.internal.concurrent.doneFuture
 import net.corda.core.internal.concurrent.openFuture
 import net.corda.core.messaging.*
+import net.corda.nodeapi.internal.pendingFlowsCount
 import net.corda.tools.shell.utlities.ANSIProgressRenderer
 import net.corda.tools.shell.utlities.StdoutANSIProgressRenderer
 import org.crsh.command.InvocationContext
@@ -42,7 +43,6 @@ import org.crsh.vfs.spi.url.ClassPathMountFactory
 import org.slf4j.LoggerFactory
 import rx.Observable
 import rx.Subscriber
-import rx.subjects.PublishSubject
 import java.io.FileDescriptor
 import java.io.FileInputStream
 import java.io.InputStream
@@ -513,36 +513,6 @@ object InteractiveShell {
             InputStreamSerializer.invokeContext = null
             InputStreamDeserializer.closeAll()
         }
-    }
-
-    // TODO sollecitom try to avoid this duplication with regards to the same function in CordaRPCOpsImpl, without exposing this as public API.
-    private fun CordaRPCOps.pendingFlowsCount(): DataFeed<Int, Pair<Int, Int>> {
-
-        val stateMachineState = stateMachinesFeed()
-        var pendingFlowsCount = stateMachineState.snapshot.size
-        var completedFlowsCount = 0
-        val updates = PublishSubject.create<Pair<Int, Int>>()
-        stateMachineState
-                .updates
-                .doOnNext { update ->
-                    when (update) {
-                        is StateMachineUpdate.Added -> {
-                            pendingFlowsCount++
-                            updates.onNext(completedFlowsCount to pendingFlowsCount)
-                        }
-                        is StateMachineUpdate.Removed -> {
-                            completedFlowsCount++
-                            updates.onNext(completedFlowsCount to pendingFlowsCount)
-                            if (completedFlowsCount == pendingFlowsCount) {
-                                updates.onCompleted()
-                            }
-                        }
-                    }
-                }.subscribe()
-        if (pendingFlowsCount == 0) {
-            updates.onCompleted()
-        }
-        return DataFeed(pendingFlowsCount, updates)
     }
 
     private fun printAndFollowRPCResponse(response: Any?, out: PrintWriter): CordaFuture<Unit> {
